requirements.txt
fastapi
uvicorn
azure-servicebus
pytest
httpx
dotenv
azure-storage-queue
